Combine PriorAuth and PriorAuthSubmit components
Remove the PriorAuth page and combine the submit component with the
claimresponse rendering component (previously called PriorAuth)
