+++ webpack.config.common.js
@@ -4,8 +4,7 @@ module.exports = {
   entry: {
     launch: path.resolve(__dirname, "src/launch.js"),
     index: path.resolve(__dirname, "src/index.js"),
-    register: path.resolve(__dirname, "src/register.js"),
-    priorauth: path.resolve(__dirname, "src/priorauth.js")
+    register: path.resolve(__dirname, "src/register.js")
   },
   output: {
     filename: "[name].bundle.js",
